+++ src/addons/outputSpokenText.js
@@ -313,6 +313,15 @@ MathAtom.toSpeakableFragment = function(atom, options) {
             case 'mclose':
             case 'textord':
             {
+                const command = atom.latex ? atom.latex.trim() : '' ;
+                if (command === '\\mathbin' || command === '\\mathrel' || 
+                    command === '\\mathopen' || command === '\\mathclose' || 
+                    command === '\\mathpunct' || command === '\\mathord' || 
+                    command === '\\mathinner') {
+                    result = MathAtom.toSpeakableFragment(atom.body, options);
+                    break;
+                }   
+
                 let atomValue = atom.body;
                 let latexValue = atom.latex;
                 if (atom.type === 'delim' || atom.type === 'sizeddelim') {